Added no functionality for users that are banned, they will not be able to log in

+++ Backend/kokottni/Data/src/main/java/coms309/Users/UserController.java
@@ -45,6 +45,7 @@ public class UserController {
 
     @GetMapping(path = "/buy/{id}/user/{uid}/amt/{amount}")
     StockPurchased purchaseStock(@PathVariable long id, @PathVariable long uid, @PathVariable int amount){
+        if(userRepository.getOne(uid).getPrivilege() == 'b') return null;
         Stock stock = stockRepository.getOne(id);
         int countBefore = userRepository.getOne(uid).getStocks().size();
         StockPurchased potentiallyRemove = userRepository.getOne(uid).purchase(amount, stock, purchaseNum);
@@ -76,19 +77,22 @@ public class UserController {
 
     @GetMapping(path = "/sell/{id}/user/{id}/{numStocks}")
     double sellStock(@PathVariable long id, @PathVariable long uid, @PathVariable int numStocks){
+        if(userRepository.getOne(uid).getPrivilege() == 'b') return -1;
         Stock stock = stockRepository.getOne(id);
         int currLength = userRepository.getOne(uid).getStocks().size();
         StockPurchased changed = userRepository.getOne(uid).sell(numStocks, stock);
+        int soldStocks = changed.getNumPurchased();
         if(currLength != userRepository.getOne(uid).getStocks().size()){
             stockPurchasedRepository.delete(changed);
         }else{
-            modifySPRepoSell(uid);
+            StockPurchased sp = modifySPRepoSell(uid);
+            soldStocks -= sp.getNumPurchased();
             stockPurchasedRepository.delete(changed);
         }
-        return stock.getCurrValue() * numStocks;
+        return stock.getCurrValue() * soldStocks;
     }
 
-    private void modifySPRepoSell(long uid){
+    private StockPurchased modifySPRepoSell(long uid){
         int foundidx = -1;
         for(long i = 0; i < stockPurchasedRepository.count(); ++i){
             for(int j = 0; j < userRepository.getOne(uid).getStocks().size() && stockPurchasedRepository.getOne(i).getUser().getId().equals(uid) && stockPurchasedRepository.getOne(i).getStock().equals(userRepository.getOne(uid).getStocks().get(j).getStock()); ++j){
@@ -101,7 +105,10 @@ public class UserController {
                 break;
             }
         }
-        stockPurchasedRepository.save(userRepository.getOne(uid).getStocks().get(foundidx));
+        if(foundidx != -1){
+            return stockPurchasedRepository.save(userRepository.getOne(uid).getStocks().get(foundidx));
+        }
+        return null;
     }
 
 
@@ -118,6 +125,7 @@ public class UserController {
     @PostMapping("/login")
     String login(@RequestBody LoginAttempt login, @PathVariable Long id){
         User user = userRepository.getOne(id);
+        if(user.getPrivilege() == 'b') return failure;
         if (user != null) {
             if (user.getPassword().equals(login.getPassword())) {
                 return "Success!";
@@ -132,6 +140,7 @@ public class UserController {
     @PutMapping("/users/{id}")
     User updateUser(@PathVariable Long id, @RequestBody User request){
         User user = userRepository.getOne(id);
+        if(user.getPrivilege() == 'b') return null;
         userRepository.save(request);
         return userRepository.getOne(id);
     }
@@ -139,6 +148,7 @@ public class UserController {
     @PutMapping("/users/{userId}/stocks/{stockId}/{numPurchasing}")
     String assignStockToUser(@PathVariable Long userId, @PathVariable Long stockId, @PathVariable int numPurchasing){
         User user = userRepository.getOne(userId);
+        if(user.getPrivilege() == 'b') return failure;
         Stock stock = stockRepository.getOne(stockId);
         stock.setUser(user, numPurchasing, userId);
         user.setStock(stock, numPurchasing, stockId);
@@ -169,8 +179,9 @@ public class UserController {
         return success;
     }
 
-    @PutMapping(path = "/stocks")
-    String updateAllStocks(){
+    @PutMapping(path = "/stocks/{uid}")
+    String updateAllStocks(@PathVariable long uid){
+        if(userRepository.getOne(uid).getPrivilege() != 'a') return failure;
         stockAPI.updateAllStocks(stockRepository);
         return success;
     }
